修改交易參數

+++ lib/models/http/parameter/check_reserve_deposit.dart
@@ -28,7 +28,7 @@ class CheckReserveDeposit {
         deposit: json["deposit"],
         reserveBalance: json["reserveBalance"].toDouble(),
         tradingTime: json["tradingTime"],
-        reward: json["reward"],
+        reward: json["reward"].toDouble(),
       );
 
   Map<String, dynamic> toJson() => {
